Add Cluster version as git hash
Add the git hash of the CHIP project as Cluster Object version. This
allows us to invalidate interviewing data when we add new Cluster
objects.

Once the spec is final, that should no longer be necessary. However, we
might still need to add bug fixes to the Cluster objects. Maybe we
should store the interviewing data in a known stable format, e.g. the
Matter TLV format.