Allow to commission with code on network only
Add a parameter which instructs the SDK to search for the device only on
the network. This is useful when using the manual pairing code, which
does not contain a device specific discovery capabilties.

This is useful to prevent the SDK from searching via Bluetooth from the
Android App (which provides a manual pairing code when using the non-App
integrated Matter onboarding flow).