Make subscription attribute update handler Thread safe
The attribute update callback of subscription is called from the Matter
SDK Thread directly. Most handling needs to run in the event loop
Thread. Do a few quick checks, and then call the rest of the handler
in the event loop Thread.

This also gets rid of a couple of `loop.create_task()` calls from
the Matter SDK thread. `loop.create_task()` is not Thread safe.