Retry certification fetch if the first attempt fails
On first attempt, we create the certificate directory. Because we use
the directory as indication when the last fetch has been made, the
directory modification timestamp is right now. This means that even
if the fetching fails, subsequent Matter server startup would not retry
fetching the credentials.

Make sure we retry in this case by clearing the modification timestamp