Use SetRawAttributeUpdateCallback to support custom clusters
Update the CHIP wheels to 2024.5.1 to get the new
RawAttributeUpdateCallback.

To support subscriptions for custom cluster use this new
RawAttributeUpdateCallback instead of AttributeUpdateCallback. This
also makes sure that we update the attributes with the same objects
on initial read and subsequent subscription updates (Python types
but not Cluster specific struct types etc.).