Refactor MatterDeviceController
This commit renames the current MatterDeviceController to
MatterDeviceControllerAPI as large parts of the code deal with the
WebSocket API. A new class MatterDeviceController got introduced which
abstracts the ChipDeviceController in a nicer way. It encapsulates
the async friendly call to the SDK (in an executor for those API where
this is still necessary). It also takes care of locking per node and
manages the subscriptions.

To MatterDeviceControllerAPI it exposes a more Pythonic API to make
our code more coherent.