Use single worker to call SDK APIs with callbacks
API calls to the native side of the SDK which have global callback
functions in ChipStack are not re-entrant since they use a shared
Future object to signal the result. Currently these are all
Commissioning and the Open Commission Window APIs. They can also be
identified if they call CallAsyncWithCompleteCallback internally.
Make sure those are called from a single Thread only.

This PR is similar to #599 but only applied to the relevant API calls.