Improve Thread network diagnostics
Not all Thread devices seem to provide the Thread Network Diagnostics
Cluster. According to the Matter specification, this cluster is
optional, even for Thread devices.

This commit only parses Thread Network Diagnostics data if the cluster
is present in the device.

For completeness, also make sure that Router Eligible End Device (REED)
are classified as end devices too (since that is their current role).