Process Shutdown events
Mark the node as offline when we receive a explicit Shutdown event
from the node.

This avoids that the node remains appearing online after an update for
the full subscription report interval before realizing that the node
disappeared and subscriptions get setup again.

Most likely this helps in other cases too.