Avoid MatterNodeData captured by Subscription callback closures
The Subscription callback closures were capturing the MatterNodeData
object, which was causing the object to be kept alive even after the
node was re-interviewed. Worse even, the callbacks were still using
the old (non-reinterviewed) MatterNodeData object. This has been
observed while working on updates, and doing two updates in a row:
The first completed update causes a re-interview. The second update then
was operating on the old MatterNodeData object, causing the node
to stay unavailable.