Reintroduce ability to get IP addresses without scope
This commit reintroduces the ability to get IP addresses without
scope. This is useful in case we are really not interested in the
scope on client side e.g. because we know we run on a different
system.

There is currently not need in Home Assistant for this feature, but
rather to break it for no good reason just keep it around, it might
be useful for someone.