Improve the open commissioning window command
Check if the commissioning window is indeed open before returning
cached values. This is to prevent returning stale values when the
commissioning window got closed, either because the device got
successfully commissioned with another network or because the
commissioning window got closed due to an error in attempting to
commission. Especially the latter case leads to very bad user
experience as folks try to use the stale (non-working) commissioning
code for the next 10 minutes and get frustrated.