Update custom_clusters.py
I added a custom cluster for energy consumption data that the Neo Coolcam Matter smart plug measures. 

However the raw values for voltage and watt will need additional scaling, the voltage raw value for e.g. 227.3 V is "2273", the watt raw value for e.g. 10.1 W is "101". Current is reported in mA. Wattaccumulated is reported in Wh, however the time frame for each accumulation measurement does not seem to be constant.  

In addition I could not yet find a raw value for total energy consumption, even so the Tuya app reports such a value.