Override Liveness timeout on shutdown
Override the liveness timeout when a node announces a shutdown. This
causes the subscription logic to notice that the device went offline
much quicker, which is especially nice for battery operated devices
which have a high liveness timeout.

We use 5s as timeout to make sure the device really completed shutdown,
and that the first subscription resumption triggered by this actually
fails. We can restore the timeout to default immediately after, since
the SDK is in subscription resumption mode now.

Another alternative would be to shutdown the subscription (mark the
device as offline). This has the disadvantage that the device needs
to recreate the subscription from scratch when it comes back, which
causes unnecessary traffic.

If the node stays offline for longer than 30s, we'll still trigger
the offline logic which will shutdown the subscription entirely.

This is mainly useful for updates, where the node goes offline
gracefully (announced with the shutdown event) and comes back soon
after.