Don't use threadsafe variant in signal_event
All calls to signal_event are made in the asyncio event loop thread.
Therefore, we don't need to use the threadsafe variant of call_soon()
anymore.

This was anyways a bit brittle as asyncio.create_task() is not
threadsafe, so calling signal_event from a different thread was not
really safe.