Raise InvalidServerVersion when Server is too old too
Currently InvalidServerVersion was only raised when the client was
too old (Server wasn't able to handle such an old client anymore),
e.g. after a backwards incompatible change Server side.

However, if the client was newer than the Server schema version, no
error was raised. This essentially never triggered the add-on update
logic in Core.

This change makes sure that we raise InvalidServerVersion when the
Server is too old for the client too. We also add properties to the
Exception so a client can figure out which situation we are dealing
with (too old/too new) and act accordingly.