Add note about Docker default command
Our Dockerfile uses the CMD instruction to run the server with sensible
default arguments. Often folks forget to add those when customizing
their command. This adds a note to the README.md and the compose.yml
file to remind them.

This also removes the D-Bus access in the default startup example since
D-Bus is only required for Bluetooth support (which is no longer
enabled by default). An additional example with the correct command
line and D-Bus access should make it clear what needs to be done to
restore previous behavior.