Check if setup task is actually still running
On first setup, we create a task and the task will remain in the task
list. If a node gets offline and rediscovered on mDNS again, we will
need to setup the node again. In this case we'll have that previous
task, but it will be done. So check if the task is done and if so allow
starting another setup task.