Set fabric label on interview
Instead of setting the label on commissioning only, set it during the
interview process but only if necessary. This avoids the commissioning
process to fail in a incomplete state when setting the label fails.

The interview process has more resilience to recover from a failure
during the commissioning since we have a retry mechanism in place.

This also allows to set the label on devices that are already
commissioned.

Also (try to) unpair the device in case interviewing fails. This allows
user to commission a second time without running into "Trying to add a
NOC for a fabric that already exists" errors.