Improve mDNS browsing to find operational nodes
Use the fabric specific DNS-SD service subtype instead of the generic
"_matter._tcp" service type. This avoids querying for operational nodes
of other fabrics. This is closer to the recommendations in
"4.3.2.6. Performance Recommendations" of the Matter spec.

While at it, remove the commissionable node browsing as it is not
used currently.