Always use multicasts for service discovery
The unicast service discovery leads to unicast responses. However, only
a single process can receive a unicast response. This is a problem
since the Matter Server is most likely not the only mDNS client on a
given host.

By default, the service discovery starts with an unicast request first
and continues with multicast requests from then onwards (the first
multicast request is sent about 200ms later). So service discovery
worked with the default initialization too. But with this change we
get rid of the unicast request speeding up the service discovery
slightly.